refactor: split dashboard.go into focused modules
Break the ~999-line dashboard.go into four focused files:

- dashboard.go: model struct, Init/Update/View, cobra command
- dashboard_commands.go: tea.Cmd factories, message types, prStatus
- dashboard_render.go: rendering helpers, styles, formatting utils
- dashboard_pipeline.go: pipeline/data helpers, repo grouping, state queries

All files remain in package cmd. No behavioral changes.

Closes #200
