+++ internal/cmd/dashboard.go
@@ -8,19 +8,16 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 	"github.com/fsnotify/fsnotify"
 	"github.com/patflynn/klaus/internal/config"
 	"github.com/patflynn/klaus/internal/event"
 	"github.com/patflynn/klaus/internal/git"
 	gh "github.com/patflynn/klaus/internal/github"
 	"github.com/patflynn/klaus/internal/pipeline"
-	"github.com/patflynn/klaus/internal/project"
 	"github.com/patflynn/klaus/internal/run"
 	"github.com/patflynn/klaus/internal/webhook"
 	"github.com/spf13/cobra"
@@ -101,58 +98,6 @@ Keyboard shortcuts:
 
 func init() {
 	rootCmd.AddCommand(dashboardCmd)
-}
-
-// Messages for the bubbletea event loop.
-
-type statesLoadedMsg struct {
-	states []*run.State
-}
-
-type ghStatusMsg struct {
-	statuses map[string]*prStatus
-}
-
-type fsEventMsg struct{}
-
-type tickMsg struct{}
-
-type sandboxStatusMsg struct {
-	hosts map[string]bool // host -> reachable
-}
-
-type pipelineActionMsg struct {
-	actions []pipeline.Action
-}
-
-type errMsg struct {
-	err error
-}
-
-type webhookMsg struct {
-	event webhook.Event
-}
-
-type trustedCommentsMsg struct {
-	prNumber             string
-	hasNewTrustedComments bool
-}
-
-// prStatus holds the GitHub-fetched status for a single PR.
-type prStatus struct {
-	PRNumber              string
-	State                 string // OPEN, MERGED, CLOSED
-	CI                    string // passing, failing, pending, unknown
-	Conflicts             string // yes, none, unknown
-	ReviewDecision        string // APPROVED, CHANGES_REQUESTED, etc.
-	HasNewTrustedComments bool   // unaddressed comments from trusted reviewers
-}
-
-// repoGroup is a set of runs and PRs belonging to the same repository.
-type repoGroup struct {
-	Repo  string
-	Runs  []*run.State
-	PRMap map[string]*prStatus
 }
 
 // dashboardModel is the bubbletea model for the dashboard.
@@ -471,528 +416,4 @@ func (m dashboardModel) View() string {
 	b.WriteString("\n")
 
 	return b.String()
-}
-
-func (m dashboardModel) renderGroup(g repoGroup) string {
-	var b strings.Builder
-
-	// Repo header with counts
-	prCount := 0
-	agentCount := 0
-	for _, s := range g.Runs {
-		if s.Type == "session" {
-			continue
-		}
-		agentCount++
-		if s.PRURL != nil {
-			prCount++
-		}
-	}
-
-	repoLabel := repoStyle.Render(fmt.Sprintf(" %s", g.Repo))
-	counts := dimStyle.Render(fmt.Sprintf(
-		"%d agent%s, %d PR%s",
-		agentCount, pluralS(agentCount),
-		prCount, pluralS(prCount),
-	))
-	b.WriteString(fmt.Sprintf("%s%s\n", repoLabel, rightAlignPad(counts, m.width-lipgloss.Width(repoLabel))))
-	b.WriteString(dimStyle.Render(" " + strings.Repeat("─", clamp(m.width-2, 0, 120))))
-	b.WriteString("\n")
-
-	// Group agents by PR number in a single pass (O(N)).
-	prToAgents := make(map[string][]*run.State)
-	var bareAgents []*run.State
-	var prOrder []string
-	seenPRs := make(map[string]bool)
-
-	for _, s := range g.Runs {
-		if s.Type == "session" {
-			continue
-		}
-		prNum := extractPRNumber(s)
-		if prNum != "" {
-			prToAgents[prNum] = append(prToAgents[prNum], s)
-			if !seenPRs[prNum] {
-				prOrder = append(prOrder, prNum)
-				seenPRs[prNum] = true
-			}
-		} else {
-			bareAgents = append(bareAgents, s)
-		}
-	}
-
-	// Render PRs and their agents
-	for _, prNum := range prOrder {
-		agents := prToAgents[prNum]
-		if len(agents) == 0 {
-			continue
-		}
-		b.WriteString(m.renderPRLine(prNum, agents, g.PRMap[prNum]))
-		b.WriteString("\n")
-		for _, s := range agents {
-			if m.isAgentRunning(s) {
-				b.WriteString(renderAgentSubline(s))
-				b.WriteString("\n")
-			}
-		}
-	}
-
-	// Render bare agents
-	for _, s := range bareAgents {
-		b.WriteString(m.renderBareAgentLine(s))
-		b.WriteString("\n")
-	}
-
-	return b.String()
-}
-
-func (m dashboardModel) renderPRLine(prNum string, agents []*run.State, ps *prStatus) string {
-	s := agents[0]
-	prLabel := fmt.Sprintf("  #%-5s", prNum)
-	prompt := truncate(s.Prompt, 20)
-
-	state := "OPEN"
-	if s.MergedAt != nil {
-		state = "MERGED"
-	} else if ps != nil && ps.State != "" {
-		state = ps.State
-	}
-
-	var parts []string
-	parts = append(parts, stateLabel(state))
-
-	if ps != nil && state == "OPEN" {
-		parts = append(parts, ciLabel(ps.CI))
-		if ps.Conflicts == "yes" {
-			parts = append(parts, redStyle.Render("conflicts ✗"))
-		}
-		rd := ps.ReviewDecision
-		if strings.EqualFold(rd, "APPROVED") {
-			parts = append(parts, greenStyle.Render("ready"))
-		} else if strings.EqualFold(rd, "CHANGES_REQUESTED") {
-			parts = append(parts, redStyle.Render("changes requested"))
-		}
-	}
-
-	// Show klaus-internal approval if any run for this PR is approved.
-	if state == "OPEN" && isAnyRunApproved(agents) {
-		parts = append(parts, cyanStyle.Render("✓ approved"))
-	}
-
-	// Append pipeline stage if available.
-	if pps, ok := m.pipelineStates[prNum]; ok {
-		parts = append(parts, dimStyle.Render(pipeline.StageLabel(pps.Stage)))
-	}
-
-	return fmt.Sprintf("%s  %-20s  %s", prLabel, prompt, strings.Join(parts, "  "))
-}
-
-// isAnyRunApproved returns true if any of the given run states has been
-// approved via `klaus approve`.
-func isAnyRunApproved(states []*run.State) bool {
-	for _, s := range states {
-		if s.Approved != nil && *s.Approved {
-			return true
-		}
-	}
-	return false
-}
-
-func renderAgentSubline(s *run.State) string {
-	shortID := shortRunID(s.ID)
-	prompt := truncate(s.Prompt, 20)
-	hostTag := sandboxTag(s)
-	return yellowStyle.Render(fmt.Sprintf("   └─ agent:%s %s...%s", shortID, prompt, hostTag))
-}
-
-func (m *dashboardModel) renderBareAgentLine(s *run.State) string {
-	shortID := shortRunID(s.ID)
-	status := agentStatusLabel(s)
-	cost := formatCost(s)
-	prompt := truncate(s.Prompt, 20)
-	hostTag := sandboxTag(s)
-
-	if m.isAgentRunning(s) {
-		return yellowStyle.Render(fmt.Sprintf("  agent:%s  %-20s  RUNNING   %s", shortID, prompt, cost)) + hostTag
-	}
-	return dimStyle.Render(fmt.Sprintf("  agent:%s  %-20s  %s   %s", shortID, prompt, status, cost)) + hostTag
-}
-
-// sandboxTag returns a styled "[sandbox]" tag if the agent ran on a sandbox host.
-func sandboxTag(s *run.State) string {
-	if s.Host != nil {
-		return " " + sandboxStyle.Render("[sandbox]")
-	}
-	return ""
-}
-
-// Commands for the bubbletea event loop.
-
-func loadStatesCmd(store run.StateStore) tea.Cmd {
-	return func() tea.Msg {
-		states, err := store.List()
-		if err != nil {
-			return errMsg{err: err}
-		}
-		return statesLoadedMsg{states: states}
-	}
-}
-
-func fetchGHStatusCmd(client gh.Client, states []*run.State) tea.Cmd {
-	return func() tea.Msg {
-		statuses := make(map[string]*prStatus)
-		seen := make(map[string]bool)
-		for _, s := range states {
-			prNum := extractPRNumber(s)
-			prRef := extractPRRef(s)
-			if prNum == "" || seen[prNum] {
-				continue
-			}
-			seen[prNum] = true
-			statuses[prNum] = fetchPRStatus(client, prNum, prRef)
-		}
-		return ghStatusMsg{statuses: statuses}
-	}
-}
-
-func waitForWebhookCmd(ch <-chan webhook.Event) tea.Cmd {
-	if ch == nil {
-		return nil
-	}
-	return func() tea.Msg {
-		ev, ok := <-ch
-		if !ok {
-			return nil
-		}
-		return webhookMsg{event: ev}
-	}
-}
-
-func tickCmd() tea.Cmd {
-	return func() tea.Msg {
-		return tickMsg{}
-	}
-}
-
-func tickAfterCmd() tea.Cmd {
-	return tea.Tick(30*time.Second, func(time.Time) tea.Msg {
-		return tickMsg{}
-	})
-}
-
-func startWatcherCmd(store run.StateStore) tea.Cmd {
-	return func() tea.Msg {
-		stateDir := store.StateDir()
-		// Ensure the directory exists before watching
-		os.MkdirAll(stateDir, 0o755)
-
-		w, err := fsnotify.NewWatcher()
-		if err != nil {
-			return errMsg{err: fmt.Errorf("creating file watcher: %w", err)}
-		}
-		if err := w.Add(stateDir); err != nil {
-			w.Close()
-			return errMsg{err: fmt.Errorf("watching state dir: %w", err)}
-		}
-		return w
-	}
-}
-
-func watchFSCmd(w *fsnotify.Watcher) tea.Cmd {
-	return func() tea.Msg {
-		for {
-			select {
-			case event, ok := <-w.Events:
-				if !ok {
-					return nil
-				}
-				if filepath.Ext(event.Name) == ".json" {
-					return fsEventMsg{}
-				}
-			case _, ok := <-w.Errors:
-				if !ok {
-					return nil
-				}
-				// Ignore watcher errors and keep watching
-			}
-		}
-	}
-}
-
-func checkSandboxCmd(states []*run.State) tea.Cmd {
-	return func() tea.Msg {
-		hosts := make(map[string]bool)
-		for _, s := range states {
-			if s.Host != nil && *s.Host != "" {
-				if _, ok := hosts[*s.Host]; !ok {
-					hosts[*s.Host] = CheckSandboxReachable(*s.Host)
-				}
-			}
-		}
-		if len(hosts) == 0 {
-			return nil
-		}
-		return sandboxStatusMsg{hosts: hosts}
-	}
-}
-
-func renderSandboxStatus(hosts map[string]bool) string {
-	if len(hosts) == 0 {
-		return ""
-	}
-	var parts []string
-	for host, reachable := range hosts {
-		if reachable {
-			parts = append(parts, greenStyle.Render(fmt.Sprintf("  sandbox %s: ✓", host)))
-		} else {
-			parts = append(parts, redStyle.Render(fmt.Sprintf("  sandbox %s: ✗", host)))
-		}
-	}
-	sort.Strings(parts)
-	return strings.Join(parts, "  ") + "\n"
-}
-
-// fetchPRStatus queries GitHub for a single PR's status.
-// prRef should be a full PR URL so gh can resolve it from any directory.
-func fetchPRStatus(client gh.Client, prNumber, prRef string) *prStatus {
-	ctx := context.TODO()
-	ps := &prStatus{PRNumber: prNumber}
-	ps.State = client.GetState(ctx, prRef)
-	if ps.State == "MERGED" || ps.State == "CLOSED" {
-		return ps
-	}
-	ps.CI = client.GetCI(ctx, prRef)
-	ps.Conflicts = client.GetConflicts(ctx, prRef)
-	ps.ReviewDecision = client.GetReviewDecision(ctx, prRef)
-
-	// When reviewDecision is not CHANGES_REQUESTED, check for unaddressed
-	// trusted reviewer comments that GitHub doesn't reflect in reviewDecision.
-	if !strings.EqualFold(ps.ReviewDecision, "CHANGES_REQUESTED") &&
-		!strings.EqualFold(ps.ReviewDecision, "APPROVED") {
-		ownerRepo := ownerRepoFromPRURL(prRef)
-		if ownerRepo != "" {
-			ps.HasNewTrustedComments = hasUnaddressedTrustedComments(ownerRepo, prNumber)
-		}
-	}
-	return ps
-}
-
-// Data layer functions (testable).
-
-// groupByRepo organizes states into repo groups, sorted by repo name.
-func groupByRepo(states []*run.State) []repoGroup {
-	reg, _ := project.Load()
-	groups := make(map[string]*repoGroup)
-	for _, s := range states {
-		if s.Type == "session" {
-			continue
-		}
-		repo := repoFromState(s, reg)
-		g, ok := groups[repo]
-		if !ok {
-			g = &repoGroup{
-				Repo:  repo,
-				PRMap: make(map[string]*prStatus),
-			}
-			groups[repo] = g
-		}
-		g.Runs = append(g.Runs, s)
-	}
-
-	var result []repoGroup
-	for _, g := range groups {
-		result = append(result, *g)
-	}
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].Repo < result[j].Repo
-	})
-	return result
-}
-
-// repoFromState extracts the repo identifier from a run state, normalizing
-// against the project registry so that different forms of the same repo
-// (e.g. "cosmo" vs "patflynn/cosmo") group together.
-func repoFromState(s *run.State, reg *project.Registry) string {
-	if s.TargetRepo != nil && *s.TargetRepo != "" {
-		return project.NormalizeRepoName(*s.TargetRepo, reg)
-	}
-	if s.PRURL != nil {
-		ownerRepo := repoFromPRURL(*s.PRURL)
-		if ownerRepo != "(unknown)" {
-			return project.NormalizeRepoName(ownerRepo, reg)
-		}
-		return ownerRepo
-	}
-	return "(local)"
-}
-
-// repoFromPRURL extracts "owner/repo" from a GitHub PR URL.
-func repoFromPRURL(prURL string) string {
-	// https://github.com/owner/repo/pull/123
-	prURL = strings.TrimPrefix(prURL, "https://github.com/")
-	prURL = strings.TrimPrefix(prURL, "http://github.com/")
-	parts := strings.Split(prURL, "/")
-	if len(parts) >= 2 {
-		return parts[0] + "/" + parts[1]
-	}
-	return "(unknown)"
-}
-
-// computeTotalCost sums the cost across all runs. For runs without final cost,
-// it includes the budget as an upper bound.
-func computeTotalCost(states []*run.State) float64 {
-	var total float64
-	for _, s := range states {
-		if s.CostUSD != nil {
-			total += *s.CostUSD
-		}
-	}
-	return total
-}
-
-// countAgents returns (running, total) agent counts (excludes sessions).
-func (m *dashboardModel) countAgents(states []*run.State) (int, int) {
-	var running, total int
-	for _, s := range states {
-		if s.Type == "session" {
-			continue
-		}
-		total++
-		if m.isAgentRunning(s) {
-			running++
-		}
-	}
-	return running, total
-}
-
-// computeSessionDuration returns the duration from the oldest active run's creation time to now.
-func computeSessionDuration(states []*run.State) time.Duration {
-	if len(states) == 0 {
-		return 0
-	}
-	var oldest time.Time
-	for _, s := range states {
-		t, err := time.Parse(time.RFC3339, s.CreatedAt)
-		if err != nil {
-			continue
-		}
-		if oldest.IsZero() || t.Before(oldest) {
-			oldest = t
-		}
-	}
-	if oldest.IsZero() {
-		return 0
-	}
-	return time.Since(oldest)
-}
-
-// markRunFailed sets sentinel finalization values on a stale run and
-// cleans up its worktree and branch. Errors during cleanup are logged
-// but do not prevent the state from being saved.
-func markRunFailed(store run.StateStore, s *run.State) {
-	cost := float64(-1)
-	dur := int64(0)
-	s.CostUSD = &cost
-	s.DurationMS = &dur
-	s.TmuxPane = nil
-	cleanupWorktree(context.Background(), store, git.NewExecClient(), s)
-	if err := store.Save(s); err != nil {
-		slog.Warn("failed to save stale run state", "id", s.ID, "err", err)
-	}
-}
-
-// isAgentRunning checks if a run's agent is currently active in tmux.
-func (m *dashboardModel) isAgentRunning(s *run.State) bool {
-	return s.IsAgentRunningWith(m.tmuxDeps)
-}
-
-// agentStatusLabel returns a display label for a non-running agent.
-func agentStatusLabel(s *run.State) string {
-	if s.PRURL != nil {
-		return "PR"
-	}
-	return "EXITED"
-}
-
-// shortRunID returns the last 4 chars of a run ID.
-func shortRunID(id string) string {
-	if len(id) < 4 {
-		return id
-	}
-	return id[len(id)-4:]
-}
-
-// formatDuration renders a duration as "Xh Ym" or "Xm Ys".
-func formatDuration(d time.Duration) string {
-	d = d.Round(time.Minute)
-	h := int(d.Hours())
-	m := int(d.Minutes()) % 60
-	if h > 0 {
-		return fmt.Sprintf("%dh %02dm", h, m)
-	}
-	return fmt.Sprintf("%dm", m)
-}
-
-// Styling.
-
-var (
-	headerStyle  = lipgloss.NewStyle().Bold(true)
-	repoStyle    = lipgloss.NewStyle().Bold(true)
-	greenStyle   = lipgloss.NewStyle().Foreground(lipgloss.Color("2"))
-	redStyle     = lipgloss.NewStyle().Foreground(lipgloss.Color("1"))
-	yellowStyle  = lipgloss.NewStyle().Foreground(lipgloss.Color("3"))
-	dimStyle     = lipgloss.NewStyle().Foreground(lipgloss.Color("8"))
-	cyanStyle    = lipgloss.NewStyle().Foreground(lipgloss.Color("6"))
-	sandboxStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("6"))
-	dimRedStyle  = lipgloss.NewStyle().Foreground(lipgloss.Color("9")).Faint(true)
-)
-
-func stateLabel(state string) string {
-	switch state {
-	case "MERGED":
-		return greenStyle.Render("MERGED")
-	case "CLOSED":
-		return dimStyle.Render("CLOSED")
-	default:
-		return yellowStyle.Render("OPEN")
-	}
-}
-
-func ciLabel(ci string) string {
-	switch ci {
-	case "passing":
-		return greenStyle.Render("CI ✓")
-	case "failing":
-		return redStyle.Render("CI ✗")
-	case "pending":
-		return yellowStyle.Render("CI …")
-	default:
-		return dimStyle.Render("CI ?")
-	}
-}
-
-func rightAlignPad(s string, totalWidth int) string {
-	w := lipgloss.Width(s)
-	pad := totalWidth - w
-	if pad <= 0 {
-		return s
-	}
-	return strings.Repeat(" ", pad) + s
-}
-
-func clamp(v, lo, hi int) int {
-	if v < lo {
-		return lo
-	}
-	if v > hi {
-		return hi
-	}
-	return v
-}
-
-func pluralS(n int) string {
-	if n == 1 {
-		return ""
-	}
-	return "s"
 }
